Test data cap against local SemApps/Jena TDB.
This is primarily based on the SemApps Docker image for Apache Jena
Fuseki with WebACL, used in the SemApps TripleStoreService:

https://semapps.org/docs/triplestore/

The SemApps LDP and API Gateway based on Molecular microservices are
not really critical to this implementation, and the WebACL has not been
configured. The Docker image just provides a good solid foundation for
persisting the JSON-LD that is exported by the Connector.